send packets in bursts, not individually

+++ src/mocking/mock_data.hpp
@@ -5,6 +5,6 @@
 
 namespace netbook::mocking {
 
-void mock_data_pusher(std::uint8_t queue_id, std::uint64_t packets_to_push);
+void mock_data_pusher(std::uint8_t queue_id, std::uint32_t packets_to_push);
 
 }
